Factored pre-processing code out into a new controller class
