+++ client/src/main/resources/views/workspace/ProcessDataView.fxml
@@ -15,7 +15,7 @@
 <Tab text="   Process Data" closable="false"
      xmlns="http://javafx.com/javafx"
      xmlns:fx="http://javafx.com/fxml"
-     fx:controller="org.iconic.workspace.WorkspaceController">
+     fx:controller="org.iconic.project.processing.ProcessDataController">
 	 
     <fx:define>
         <Screen fx:factory="getPrimary" fx:id="screen"/>
